split up components into components or elements
