+++ src/index.js
@@ -7,8 +7,9 @@ import '@tailwindcss/forms/dist/forms.css';
 import { Store, create } from 'microstates';
 import AppModel, { State } from './state';
 
-import Header from './components/common/header';
-import Footer from './components/common/footer';
+import Header from './elements/Header';
+import Footer from './elements/Footer';
+
 import Homepage from './pages/homepage';
 import Examples from './pages/examples';
 import Financial from './pages/flow';
